♻️ Refactor connectWallet functions into mixin
